Add MaxNodesInliner to inline calls until graph reaches certain size
