+++ src/main/java/minijava/ir/optimize/Inliner.java
@@ -29,13 +29,13 @@ import minijava.ir.utils.GraphUtils;
 import minijava.ir.utils.NodeUtils;
 import org.jooq.lambda.tuple.Tuple2;
 
-public class Inliner extends BaseOptimizer {
+abstract class Inliner extends BaseOptimizer {
   private static final Set<ir_opcode> ALWAYS_IN_START_BLOCK =
       Sets.newHashSet(iro_Const, iro_Address);
-  private final ProgramMetrics metrics;
-  private final Set<Call> callsToInline = new HashSet<>();
+  final ProgramMetrics metrics;
+  final Set<Call> callsToInline = new HashSet<>();
 
-  public Inliner(ProgramMetrics metrics) {
+  Inliner(ProgramMetrics metrics) {
     this.metrics = metrics;
   }
 
@@ -43,21 +43,14 @@ public class Inliner extends BaseOptimizer {
   public boolean optimize(Graph graph) {
     this.graph = graph;
     this.callsToInline.clear();
+    this.metrics.updateGraphInfo(graph);
     fixedPointIteration();
     return inlineCandidates();
   }
 
-  private boolean inlineCandidates() {
-    boolean hasChanged = false;
-    for (Call call : callsToInline) {
-      //System.out.println("Inlining " + callee(call) + " into " + graph);
-      inline(call);
-      hasChanged = true;
-    }
-    return hasChanged;
-  }
+  abstract boolean inlineCandidates();
 
-  private void inline(Call call) {
+  void inline(Call call) {
     Graph callee = callee(call);
     Tuple2<Start, End> subgraph = GraphUtils.copyGraph(callee, call.getGraph());
     Start start = subgraph.v1;
@@ -212,33 +205,7 @@ public class Inliner extends BaseOptimizer {
         });
   }
 
-  @Override
-  public void visit(Call call) {
-    Graph callee = callee(call);
-    if (callee == null) {
-      // This was a foreign/native call, where we don't have access to the graph.
-      return;
-    }
-    ProgramMetrics.GraphInfo calleeInfo = metrics.graphInfos.get(callee(call));
-    if (!calleeInfo.calls.isEmpty()) {
-      // We only inline if the callee itself doesn't call any other functions for now
-      return;
-    }
-    if (calleeInfo.diverges) {
-      // We could potentially inline this, but it won't bring any benefit, as the diverging loop
-      // dominates the method call overhead.
-      // In case we do want this though, we have to graph.keepAlive(end.getPred(1));
-      return;
-    }
-    if (calleeInfo.size > 1000) {
-      // This should be sufficiently high so that the call overhead isn't noticeable
-      // (generally, imagine an if/else...).
-      return;
-    }
-    callsToInline.add(call);
-  }
-
-  private Graph callee(Call call) {
+  Graph callee(Call call) {
     Address funcPtr = (Address) call.getPtr();
     return funcPtr.getEntity().getGraph();
   }
